amplifier-ado-event-monitor/pyproject.toml
[project]
name = "ado-event-monitor"
version = "0.1.0"
description = "Autonomous event monitor for Azure DevOps - detects changes and triggers Amplifier agents"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "pyright>=1.1",
]

[project.scripts]
ado-monitor = "ado_monitor.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ado_monitor"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "C4", "SIM"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
